+++ src/remove_colors/mod.rs
@@ -1,3 +1,0 @@
-pub mod remove_colors;
-
-pub use remove_colors::remove_ansi_colors;