+++ src/weather/mod.rs
@@ -1,6 +1,6 @@
 pub mod weather;
 mod fetch_weather;
 mod get_ascii_art;
-mod fmt_lines;
+mod utils;
 
 pub use weather::Weather;