+++ src/weather/utils/mod.rs
@@ -1,0 +1,3 @@
+pub mod colorize;
+pub mod get_data;
+pub mod format;