+++ CHANGELOG.md
@@ -1,8 +1,18 @@
 # Changelog
 
+## [0.1.2]
+### Added
+- Color of UV Index
+- Name of UV Index
+### Changes
+- Big refactoring of [fetch_weather](src/weather/fetch_weather.rs)
+- Renamed names of variables
+- All functions not in Weather (struct) moved to [utils](src/weather/utils)
+- Some small changes
+
 ## [0.1.1]
 ### Added
 - Humidity 
 
-### Changes
+### Changed
 - Now printing info & ascii art works like in Neofetch